Replace unsupported/abandoned raymond package with handlebars/v3
The raymond package appears to be abandoned. This changes to a supported
fork (handlebars/v3) that has Go module support, and has some
maintenance done. Additionally, updates to latest sprig package, and
changes the logic used to skip/avoid adding sprig helpers to handlebars
when the func has an unsupported return type.